Fixed error in assigning boolean logic for TOP
Original approach was to assign the variable as True/False, but this
was creating strings, not true Booleans.  Trying to do this with
type=bool also did not work, so I went with store_action options and
changed the syntax to --top (default if not specified) and --notop.
This replaces the old syntax (that didn't work) of --top False.