FIX: ability to save 3D res_sim data
As detailed in Issue #102, changes to extract an arbitrarily-oriented 2D
plane seem to have broken the ability to extract 3D data.  This was
confirmed by rolling back `v8.2.0` to `v8.1.4`.

It appears that b/c the new `plane_*` input parameters were set to
default input parameters of 0.0/0 in `save_res()`, the `elev` vector for
3D data saving was overwritten by a single value for `elev`, which
broke things downstream.

Now, the `plane_*` input variables default to `None`, which should only
allow code that uses those values to kick in when just a single plane is
explicitly specified.