C-Cpp post-processing builds (#110)
* utilize `CMake` to build `create_disp_dat_fast` and `scale_disp_dat`; win for dynamically choosing Python install version over hard-coded `Makefile` path.
* default to Python-based `create_disp_dat.py`, but log warning about using the slow implementation and encourage using `create_disp_dat_fast`.
* unit tests clean up to keep all testing data in `tests/`
* Closes #16 

SWIG-based `*.so` generation and `cmake` compilation are still manually executed (if desired).  Tried using `scikit-build` and some other tricks to automatically compile and install the `_create_disp_dat.so` module, but it was so fragile and complicated, that it ultimately was not worth it (at this time).  This also keeps everything, by default, in a completely Python-based virtualenv workflow.